Servo Code
This code was used to rotate the servo motor a certain amount of degrees depending on the value of the flex sensor sent from the Esp-32 on the glove. If the flex sensor value was less than or equal to 475, the servo would rotate 150 degrees opening the hand. If the value of the flex sensor is over 475 degrees, the servo will rotate 20 degrees closing the hand.